% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doShard.R
\name{collect_iterations}
\alias{collect_iterations}
\title{Collect Iterations from Iterators}
\usage{
collect_iterations(it_args)
}
\arguments{
\item{it_args}{List of iterator arguments.}
}
\value{
List of iteration value lists.
}
\description{
Collects all values from foreach iterator arguments.
}
\keyword{internal}
