% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doShard.R
\name{combine_results}
\alias{combine_results}
\title{Combine Results}
\usage{
combine_results(results, accumulator, combine_info)
}
\arguments{
\item{results}{List of results.}

\item{accumulator}{Combine function.}

\item{combine_info}{Combine info from foreach.}
}
\value{
Combined result.
}
\description{
Applies the foreach combine function to results.
}
\keyword{internal}
