% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doShard.R
\docType{package}
\name{doShard}
\alias{doShard-package}
\alias{doShard}
\title{doShard: Foreach Parallel Adapter for shard}
\description{
Provides a foreach backend using the shard package for parallel execution.
}
\author{
\strong{Maintainer}: First Last \email{you@example.org}

}
