% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doShard.R
\name{doShard_backend}
\alias{doShard_backend}
\title{doShard Backend Function}
\usage{
doShard_backend(obj, expr, envir, data)
}
\arguments{
\item{obj}{The foreach object.}

\item{expr}{The expression to evaluate.}

\item{envir}{The evaluation environment.}

\item{data}{Backend options.}
}
\value{
Combined results.
}
\description{
The main backend function called by foreach to execute iterations.
}
\keyword{internal}
