% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doShard.R
\name{getDoShardPool}
\alias{getDoShardPool}
\title{Get the doShard Worker Pool}
\usage{
getDoShardPool()
}
\value{
A \code{shard_pool} object or NULL.
}
\description{
Returns the current shard worker pool, if one exists.
}
