% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doShard.R
\name{info_doShard}
\alias{info_doShard}
\title{Backend Info Function}
\usage{
info_doShard(data, item)
}
\arguments{
\item{data}{Backend options.}

\item{item}{Info item to retrieve.}
}
\value{
Requested info.
}
\description{
Returns information about the doShard backend for foreach.
}
\keyword{internal}
